move training requests out
